Small fixes + adding a separate hadron type (0) for a stop token / padding hadron
